+++ src/lib/index.ts
@@ -16,6 +16,7 @@ export * from './components/Slider/index.js';
 export * from './components/Tags/index.js';
 export * from './components/Collapsible/index.js';
 export * from './components/Tooltip/index.js';
+export * from './components/Tree/index.js';
 
 export * from './actions/portal.js';
 export * from './actions/outside.js';